requirements.txt
asgiref==3.3.4
Django==3.2.4
django-address==0.2.5
django-birthday==0.1.3
django-countries==7.2.1
django-friendship==1.9.1
django-localflavor==3.1
greenlet==1.1.0
msgpack==1.0.2
mysqlclient==2.0.3
Pillow==8.2.0
pynvim==0.4.3
python-stdnum==1.16
pytz==2021.1
sqlparse==0.4.1
